Anatoli

+++ backend/modules/deleteRestaurant.js
@@ -2,17 +2,23 @@ const deleteRestaurant = async (req, res, Restaurant, id, fs, image) => {
   try {
     const restaurant = await Restaurant.findByIdAndDelete(id);
 
-    let path = `../../frontend/public/images/${image}`;
-    fs.unlink(path, (error) => {
-      if (error) {
-        console.log(error.message);
-      } else {
-        return res.json({
-          message: "Restaurant deleted successfully",
-          status: "success",
-        });
-      }
-    });
+    if (restaurant) {
+      let path = `../../frontend/public/images/${image}`;
+      fs.unlink(path, (error) => {
+        if (error) {
+          console.log(error.message);
+        }
+      });
+      return res.json({
+        message: "Restaurant deleted successfully",
+        status: "success",
+      });
+    } else {
+      return res.json({
+        message: "Failed to delete the restaurant",
+        status: "failed",
+      });
+    }
   } catch (error) {
     return res.json({ message: error.message });
   }
